bug fix (#638)
* fix: init get collection

* fix: target host ellipsis

* fix: report colunms deps on token

* fix: deps filter change

* fix: deps filter change

* feat: replay report full api path

* refactor: split CreatePlanModal component

* feat: replay create plan protocol input

* fix: keep record case drawer after click

* feat: setting user info

* fix: diffMatch toString

* fix: update collectionsFlatData when create root collection node

---------
