+++ packages/arex/src/i18n/locales/cn/common.json
@@ -94,5 +94,6 @@
   "chromeVersionTooLow" : "当前Chrome版本过低，部分功能可能无法正常使用， 建议升级至最新版本",
   "share": "分享",
   "new": "新",
-  "more": "更多..."
+  "more": "更多...",
+  "email": "邮箱"
 }
