+++ packages/arex/src/i18n/locales/cn/components.json
@@ -89,6 +89,7 @@
     "caseEndTime": "Case 结束时间",
     "caseRange": "回放范围",
     "caseRangeTooltip": "选择用例的录制时间范围，回放该时间段内录制到的用例。",
+    "illegalProtocol": "协议不能为空或存在格式错误",
     "emptyHost": "目标回放环境不能为空",
     "emptyCaseRange": "回放范围不能为空",
     "emptyStartTime": "Case 开始时间不能为空",
@@ -125,6 +126,7 @@
     "totalCases": "Case 数",
     "planItemID": "子任务 ID",
     "api": "API",
+    "fullAPIPath": "Full API Path",
     "timeConsumed": "耗时(秒)",
     "cases": "case 数",
     "action": "操作",
@@ -432,6 +434,7 @@
     "closeRight":"关闭右侧标签页"
   },
   "systemSetting": {
+    "userInfo": "用户信息",
     "userInterface": "用户界面",
     "compact": "紧凑",
     "theme": "主题",
