+++ packages/arex/src/panes/SystemSetting/SystemSetting.tsx
@@ -2,12 +2,14 @@ import { ArexPaneFC } from '@arextest/arex-core';
 import React from 'react';
 
 import Advanced from './Advanced';
+import UserInfo from './UserInfo';
 import UserInterface from './UserInterface';
 import Version from './Version';
 
 const SystemSetting: ArexPaneFC = () => {
   return (
     <>
+      <UserInfo />
       <UserInterface />
       <Version />
       <Advanced />
