+++ packages/arex/vite.config.ts
@@ -27,9 +27,9 @@ export default defineConfig(async ({ mode }) => {
     resolve: {
       alias: {
         '@': path.resolve('./src'),
-        '@arextest/arex-core/dist': path.resolve('../arex-core/src'),
-        '@arextest/arex-core': path.resolve('../arex-core/src'),
-        '@arextest/arex-request': path.resolve('../arex-request/src'),
+        // '@arextest/arex-core/dist': path.resolve('../arex-core/src'),
+        // '@arextest/arex-core': path.resolve('../arex-core/src'),
+        // '@arextest/arex-request': path.resolve('../arex-request/src'),
       },
     },
     plugins: [
